Consult history during initial move scoring
This is certainly cleaner, and doesn't seem to be worse in some very
quick testing. May be worth a better test later, but for now I like how
clean it is.

+++ src/moveiter.h
@@ -6,7 +6,7 @@
 #include "move.h"
 #include "types.h"
 
-typedef int8_t MoveScore;
+typedef int32_t MoveScore;
 
 typedef struct {
   Movelist* list;
